Update CustomExporter.rb
Improved handling of when the resolved path of an file exceeds 260 characters.  Long paths like this are truncated into a "PATH_TOO_LONG" directory in the export directory.  Entries like this are also recorded in a file named "PATH_TOO_LONG.TXT" to help user troubleshoot what the path was too long.