add support for "metadata" stamp type
This stamp type allows any metadata to be stamped onto PDFs/TIFFs, including custom metadata.  This should help support situations where you need a per item stamped value that is tricky to get on the PDFs/TIFFs using one of the other choices, assuming you are able to get your desired value on to the item as custom metadata.